Función recuperar direcciones usuario extraída a un fichero
