+++ webapp/src/components/Navbar/Navbar.tsx
@@ -1,7 +1,6 @@
 import * as React from 'react'
 import ShoppingCartDrawer from '../Cart/ShoppingCartDrawer'
-import { CombinedDataProvider, useSession,Image } from "@inrupt/solid-ui-react";
-import { VCARD } from "@inrupt/lit-generated-vocab-common";
+import { useSession} from "@inrupt/solid-ui-react";
 
 import {
   AppBar,
@@ -57,8 +56,6 @@ const NavBar = () => {
     navigate('/home')
     setMessage('Successful logout')
   }
-
-  const { webId } = session.info as any;
 
   return (
     <>
@@ -169,16 +166,7 @@ const NavBar = () => {
           <Box sx={{ flexGrow: 0 }}>
             <Tooltip title="Open settings">
               <IconButton onClick={handleOpenUserMenu} sx={{ p: 0 }}>
-              {!session.info.isLoggedIn ? (
-                  <Avatar alt="Remy Sharp" src="/static/images/avatar/2.jpg" />
-                ):(
-                  <CombinedDataProvider datasetUrl={webId} thingUrl={webId} >
-                    <Image className="imagen" property={VCARD.hasPhoto.iri.value} 
-                      errorComponent={() => <img className='img-noPhoto' src="/images/no-image-profile.png" style={{width: '100%'}}/>}
-                      
-                    /> 
-                  </CombinedDataProvider>
-                )}
+                <Avatar alt="Remy Sharp" src="/static/images/avatar/2.jpg" />
               </IconButton>
             </Tooltip>
             <Menu
